missing casting to jint at some function calls

+++ branches/dev-1_3_3-aces/src/native/d3d/RasterRetained.cpp
@@ -213,8 +213,8 @@ void JNICALL Java_javax_media_j3d_RasterRetained_execute(JNIEnv *env,
 							 intData_array,  NULL);
 		copyDepthToSurface(d3dCtx,
 				   device, 
-				   screenCoord.sx, 
-				   screenCoord.sy, 
+				   (int)(screenCoord.sx), 
+				   (int)(screenCoord.sy), 
 				   x_offset, y_offset,
 				   w, h,width, height,
 				   intData, d3dCtx->depthStencilSurface);
@@ -230,8 +230,8 @@ void JNICALL Java_javax_media_j3d_RasterRetained_execute(JNIEnv *env,
 							      floatData_array, NULL);
 		copyDepthToSurface(d3dCtx,
 				   device, 
-				   screenCoord.sx, 
-				   screenCoord.sy, 
+				   (jint)(screenCoord.sx), 
+				   (jint)(screenCoord.sy), 
 				   x_offset, y_offset, 
 				   w, h, width, height,
 				   floatData, d3dCtx->depthStencilSurface);
